init timeout
