+++ src/prefect/concurrency/asyncio.py
@@ -1,6 +1,6 @@
 import asyncio
 from contextlib import asynccontextmanager
-from typing import List, Literal, Union, cast
+from typing import List, Literal, Optional, Union, cast
 
 import httpx
 import pendulum
@@ -13,6 +13,7 @@ except ImportError:
 
 from prefect import get_client
 from prefect.client.schemas.responses import MinimalConcurrencyLimitResponse
+from prefect.utilities.timeout import timeout_async
 
 from .events import (
     _emit_concurrency_acquisition_events,
@@ -26,21 +27,24 @@ class ConcurrencySlotAcquisitionError(Exception):
 
 
 @asynccontextmanager
-async def concurrency(names: Union[str, List[str]], occupy: int = 1):
+async def concurrency(
+    names: Union[str, List[str]], occupy: int = 1, timeout: Optional[float] = None
+):
     names = names if isinstance(names, list) else [names]
 
     limits = await _acquire_concurrency_slots(names, occupy)
     acquisition_time = pendulum.now("UTC")
     emitted_events = _emit_concurrency_acquisition_events(limits, occupy)
 
-    try:
-        yield
-    finally:
-        occupancy_period = cast(Interval, (pendulum.now("UTC") - acquisition_time))
-        await _release_concurrency_slots(
-            names, occupy, occupancy_period.total_seconds()
-        )
-        _emit_concurrency_release_events(limits, occupy, emitted_events)
+    async with timeout_async(seconds=timeout):
+        try:
+            yield
+        finally:
+            occupancy_period = cast(Interval, (pendulum.now("UTC") - acquisition_time))
+            await _release_concurrency_slots(
+                names, occupy, occupancy_period.total_seconds()
+            )
+            _emit_concurrency_release_events(limits, occupy, emitted_events)
 
 
 async def rate_limit(names: Union[str, List[str]], occupy: int = 1):
