+++ src/prefect/concurrency/sync.py
@@ -1,5 +1,5 @@
 from contextlib import contextmanager
-from typing import List, Union, cast
+from typing import List, Optional, Union, cast
 
 import pendulum
 
@@ -12,6 +12,7 @@ except ImportError:
 from prefect._internal.concurrency.api import create_call, from_sync
 from prefect._internal.concurrency.event_loop import get_running_loop
 from prefect.client.schemas.responses import MinimalConcurrencyLimitResponse
+from prefect.utilities.timeout import timeout as timeout_context
 
 from .asyncio import (
     _acquire_concurrency_slots,
@@ -24,7 +25,9 @@ from .events import (
 
 
 @contextmanager
-def concurrency(names: Union[str, List[str]], occupy: int = 1):
+def concurrency(
+    names: Union[str, List[str]], occupy: int = 1, timeout: Optional[float] = None
+):
     names = names if isinstance(names, list) else [names]
 
     limits: List[MinimalConcurrencyLimitResponse] = _call_async_function_from_sync(
@@ -33,14 +36,18 @@ def concurrency(names: Union[str, List[str]], occupy: int = 1):
     acquisition_time = pendulum.now("UTC")
     emitted_events = _emit_concurrency_acquisition_events(limits, occupy)
 
-    try:
-        yield
-    finally:
-        occupancy_period = cast(Interval, pendulum.now("UTC") - acquisition_time)
-        _call_async_function_from_sync(
-            _release_concurrency_slots, names, occupy, occupancy_period.total_seconds()
-        )
-        _emit_concurrency_release_events(limits, occupy, emitted_events)
+    with timeout_context(seconds=timeout):
+        try:
+            yield
+        finally:
+            occupancy_period = cast(Interval, pendulum.now("UTC") - acquisition_time)
+            _call_async_function_from_sync(
+                _release_concurrency_slots,
+                names,
+                occupy,
+                occupancy_period.total_seconds(),
+            )
+            _emit_concurrency_release_events(limits, occupy, emitted_events)
 
 
 def rate_limit(names: Union[str, List[str]], occupy: int = 1):
